% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.EGMME.R
\name{Change-ergmTerm}
\alias{Change-ergmTerm}
\alias{InitErgmTerm.Change (dynamic)}
\title{The Change Operator Term}
\usage{
# binary: Change(formula)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}
}
\description{
The Change Operator Term
}
\details{
This term accepts a model formula
and produces the corresponding model for a network constructed
by taking the dyads that have changed between time steps.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{durational}
\concept{operator}
