% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.godfather.R
\name{control.tergm.godfather}
\alias{control.tergm.godfather}
\title{Control parameters for \code{\link[=tergm.godfather]{tergm.godfather()}}.}
\usage{
control.tergm.godfather(GF.init.maxedges.mul = 5)
}
\arguments{
\item{GF.init.maxedges.mul}{How much space
is allocated for the edgelist of the final network. It is used
adaptively, so should not be greater than \code{10}.}
}
\description{
Returns a list of its arguments.
}
