% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.netseries.R
\name{Cross-ergmTerm}
\alias{Cross-ergmTerm}
\alias{InitErgmTerm.Cross}
\title{The Crossection Operator Term}
\usage{
# binary: Cross(formula)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}
}
\description{
The Crossection Operator Term
}
\details{
This term accepts a model formula
and produces the corresponding model for the cross-sectional
network. It is mainly useful for CMLE estimation, and has no effect (i.e.,
\code{Cross(~TERM) == ~TERM} ) for EGMME and dynamic simulation.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "Cross", "subsection")}
}
\concept{durational}
\concept{operator}
