% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.netseries.R
\name{Cross-ergmTerm}
\alias{Cross-ergmTerm}
\alias{InitErgmTerm.Cross}
\title{The Crossection Operator Term}
\usage{
# binary: Cross(formula)
}
\arguments{
\item{formula}{model formula}
}
\description{
The Crossection Operator Term
}
\details{
This term accepts a model formula
and produces the corresponding model for the cross-sectional
network. It is mainly useful for CMLE estimation, and has no effect (i.e.,
\code{Cross(~TERM) == ~TERM} ) for EGMME and dynamic simulation.
}
\concept{durational}
\concept{operator}
