% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.hints.R
\name{discord-ergmConstraint}
\alias{discord-ergmConstraint}
\alias{InitErgmConstraint.discord}
\title{The Dynamic Discordance Hint}
\usage{
# discord
}
\description{
The Dynamic Discordance Hint
}
\details{
Propose toggling discordant dyads with greater
frequency (typically about 50 percent).  May be used in
dynamic fitting and simulation.
}
