% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.hints.R
\name{discord-ergmHint}
\alias{discord-ergmHint}
\alias{InitErgmConstraint.discord}
\alias{discord-ergmConstraint}
\title{Discordant dyads}
\usage{
# discord
}
\description{
Propose toggling discordant dyads with greater
frequency (typically about 50 percent).  May be used in
dynamic fitting and simulation.
}
\seealso{
\code{\link{ergmHint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmHint", "discord", "subsection")}
}
\concept{dyad-independent}
