% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.DynMoME.R
\name{discordBDStratTNT-ergmProposal}
\alias{discordBDStratTNT-ergmProposal}
\alias{InitErgmProposal.discordBDStratTNT}
\title{Temperal TNT proposal with degree bounds}
\description{
A temporal version of \code{\link[ergm:BDStratTNT-ergmProposal]{BDStratTNT}}.  Within each
mixing type, approximately 50\\% of proposed toggles are made on
dyads, and approximately 50\\% of proposed toggles are
proposals from the network, all subject to the bounded degree
and mixing type constraints.  The degree bound constraint is imposed
the instantaneous network state
rather than the temporal operator networks).

arguments are the same as for \code{\link[ergm:BDStratTNT-ergmProposal]{BDStratTNT}},
and should be passed in via the \code{\link[ergm:bd-ergmConstraint]{bd}} and
\code{\link[ergm:blocks-ergmConstraint]{blocks}} constraints and
\code{\link[ergm:strat-ergmHint]{strat}} hint.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="discordBDStratTNT"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="discordBDStratTNT"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="discordBDStratTNT"))}}
}
\seealso{
\code{\link[ergm:ergmProposal]{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "discordBDStratTNT", "subsection")}
}
\keyword{internal}
