% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.DynMoME.R
\name{discordTNT-ergmProposal}
\alias{discordTNT-ergmProposal}
\alias{InitErgmProposal.discordTNT}
\title{Temperal TNT proposal}
\description{
A temporal version of \code{\link[ergm:TNT-ergmProposal]{TNT}}, with approximately
\code{discordance_fraction} of proposed toggles being made on the set of discordant dyads,
approximately \code{1 - discordance_fraction} of proposed toggles being TNT proposals from
network.  The value of \code{discordance_fraction} can be set by the user as a proposal argument,
defaults to \code{0.5}.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="discordTNT"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="discordTNT"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="discordTNT"))}}
}
\seealso{
\code{\link[ergm:ergmProposal]{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "discordTNT", "subsection")}
}
\keyword{internal}
