% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.EGMME.R
\name{Diss-ergmTerm}
\alias{Diss-ergmTerm}
\alias{InitErgmTerm.Diss (dynamic)}
\title{The Dissolution Operator Term}
\usage{
# binary: Diss(formula)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}
}
\description{
The Dissolution Operator Term
}
\details{
This term accepts a model formula
and produces the corresponding model for the post-dissolution
network (same as \code{Persist()} ), but with all statistics negated.

Note: This is not the equivalent of the old style \code{dissolution} model,
because the signs of the coefficients are reversed. So a larger positive
coefficient for \code{Diss()} operator means more dissolution.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "Diss", "subsection")}
}
\concept{durational}
\concept{operator}
