% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.EGMME.R
\name{Diss-ergmTerm}
\alias{Diss-ergmTerm}
\alias{InitErgmTerm.Diss (dynamic)}
\title{The Dissolution Operator Term}
\usage{
# binary: Diss(formula)
}
\arguments{
\item{formula}{model formula}
}
\description{
The Dissolution Operator Term
}
\details{
This term accepts a model formula
and produces the corresponding model for the post-dissolution
network (same as \code{Persist()} ), but with all statistics negated.

Note: This is not the equivalent of the old style \code{dissolution} model,
because the signs of the coefficients are reversed. So a larger positive
coefficient for \code{Diss()} operator means more dissolution.
}
\concept{durational}
\concept{operator}
