% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.duration.R
\name{edge.ages-ergmTerm}
\alias{edge.ages-ergmTerm}
\alias{InitErgmTerm.edge.ages}
\title{Sum of ages of extant ties}
\usage{
# binary: edge.ages
}
\description{
Sum of ages of extant ties
}
\details{
This term adds one statistic equaling sum, over all ties
present in the network, of the amount of time elapsed since
formation.

Unlike \code{\link[=mean.age-ergmTerm]{mean.age}} , this statistic is well-defined on
an empty network. However, if used as a target, it appears to
produce highly biased dissolution parameter estimates if the goal
is to get an intended average duration.
}
