% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.duration.R
\name{EdgeAges-ergmTerm}
\alias{EdgeAges-ergmTerm}
\alias{InitErgmTerm.EdgeAges}
\title{The EdgeAges Operator Term}
\usage{
# binary: EdgeAges(formula)
}
\arguments{
\item{formula}{cross-sectional, dyad-independent model formula}
}
\description{
The EdgeAges Operator Term
}
\details{
This term accepts a cross-sectional, dyad-independent model
formula.  The statistics of the EdgeAges term are equal to the sum over
all extant ties of the tie age times the on-toggle change statistics for
the tie under the given model formula.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "EdgeAges", "subsection")}
}
\concept{durational}
\concept{operator}
