% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.duration.R
\name{edges.ageinterval-ergmTerm}
\alias{edges.ageinterval-ergmTerm}
\alias{InitErgmTerm.edges.ageinterval}
\title{Number of edges with age falling into a specified range}
\usage{
# binary: edges.ageinterval(from, to=+Inf)
}
\arguments{
\item{from, to}{parameters to specify the lower bound and strict upper bounds. Can be scalars, vectors of the same length, or one of them must have length one, in which case it is recycled.}
}
\description{
Number of edges with age falling into a specified range
}
\details{
This term counts the number of edges in the network for
which the time elapsed since formation is greater than or equal to
\code{from} but strictly less than \code{to} . In other words, it
is in the semiopen interval \verb{[from, to)} .
}
