% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm-terms-index.R
\docType{package}
\name{ergmConstraint}
\alias{ergmConstraint}
\alias{ergm-constraints}
\alias{constraints-ergm}
\alias{ergm.constraints}
\alias{constraints.ergm}
\alias{ergm-hints}
\alias{hints-ergm}
\alias{ergm.hints}
\alias{hints.ergm}
\title{Constraints and hints for Temporal Exponential Family Random Graph Models}
\description{
This page describes the hints and network sample
space constraints that are included with the \code{\link[=tergm-package]{tergm}}
package. For more information, and instructions for using hints and constraints, see
\code{\link[ergm:ergmConstraint]{ergmConstraint}} and \code{\link{ergm}}.
}
\section{Constraints and hints to represent network statistics included in the \code{\link[=tergm-package]{tergm}} package}{

\subsection{Term index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmConstraint"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmConstraint"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmConstraint"))}}
}
}

\references{
\itemize{
\item Krivitsky PN and Handcock MS (2014) A Separable Model for Dynamic Networks.
\emph{Journal of the Royal Statistical Society, Series B}, 76(1): 29-46. \doi{10.1111/rssb.12014}
\item Goodreau SM, Handcock MS, Hunter DR, Butts CT, Morris M (2008a).  A \pkg{statnet} Tutorial.
\emph{Journal of Statistical Software}, 24(8).  \url{https://www.jstatsoft.org/v24/i08/}.
\item Hunter, D. R. and Handcock, M. S. (2006) Inference in curved exponential family models for networks,
\emph{Journal of Computational and Graphical Statistics}.
\item Hunter DR, Handcock MS, Butts CT, Goodreau SM, Morris M (2008b).
\pkg{ergm}: A Package to Fit, Simulate and Diagnose Exponential-Family Models for Networks.
\emph{Journal of Statistical Software}, 24(3). \url{https://www.jstatsoft.org/v24/i03/}.
\item Krivitsky PN (2012). Exponential-Family Random Graph Models for Valued Networks.
\emph{Electronic Journal of Statistics}, 2012, 6, 1100-1128. \doi{10.1214/12-EJS696}
\item Morris M, Handcock MS, Hunter DR (2008).  Specification of Exponential-Family Random Graph Models:
Terms and Computational Aspects. \emph{Journal of Statistical Software}, 24(4).
\url{https://www.jstatsoft.org/v24/i04/}.
}
}
\keyword{models}
