% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm-terms-index.R
\docType{package}
\name{ergmTerm}
\alias{ergmTerm}
\alias{ergm-terms}
\alias{ergm.terms}
\alias{terms-ergm}
\alias{terms.ergm}
\alias{nodemix.mean.age}
\alias{InitErgmTerm.edgecov}
\title{Temporally-Sensitive Operator and Durational Terms used in Exponential
Family Random Graph Models}
\description{
The terms described here are unique to temporal networks: each summarizes
some type of change or durational information.

The operator terms include: \code{Form()}, \code{Persist()}, \code{Diss()},
\code{Cross()} and \code{Change()}.  These are used to specify how the
\code{\link[ergm:ergm-terms]{ergm-terms}} in a formula are evaluated across a network
time-series.  Note, you cannot use an operator within another operator, so
\code{Cross(~Form(~edges))} is not a valid specification.

The durational terms are distinguished either by their name,
\code{mean.age}, or their name extensions: \verb{<name>.ages},
\verb{<name>.mean.age}, and \verb{<name>.age.interval}.  In contrast to
their named \code{\link[ergm:ergm-terms]{ergm-terms}} counterparts, these durational
terms take into account the elapsed time since each (term-relevant) dyad in
the network was last toggled.

As currently implemented, the package does not support use of durational
terms during estimation.  But durational terms may be used as targets,
monitors, or summary statistics.  The ability to use these terms in the
estimation of models is under development.
}
\section{Terms to represent network statistics included in the \code{\link[=tergm-package]{tergm}} package}{

\subsection{Term index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmTerm"))}}
}

\subsection{Frequently-used terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
}

\subsection{Operator terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
}

\subsection{All terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm"))}}
}

\subsection{Terms by concept}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocLatex(ergm:::.termToc("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocText(ergm:::.termToc("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocHtml(ergm:::.termToc("ergmTerm"))}}
}
}

\references{
\itemize{
\item Krivitsky, P.N. (2012). Modeling of Dynamic Networks based on
Egocentric Data with Durational Information. \emph{Pennsylvania State
University Department of Statistics Technical Report}, 2012(2012-01).
\url{https://web.archive.org/web/20170830053722/https://stat.psu.edu/research/technical-report-files/2012-technical-reports/TR1201A.pdf}
\item Krivitsky, P.N. (2012). Modeling Tie Duration in ERGM-Based Dynamic
Network Models. \emph{Pennsylvania State University Department of Statistics
Technical Report}, 2012(2012-02).
}
}
\seealso{
\code{\link[ergm:ergm-terms]{ergm-terms}} (from the
\code{\link[=ergm-package]{ergm}} package), \code{\link{ergm}},
\code{\link{network}}, \code{\link{\%v\%}}, \code{\link{\%n\%}}
}
\keyword{models}
