% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.EGMME.R
\name{Form-ergmTerm}
\alias{Form-ergmTerm}
\alias{InitErgmTerm.Form (dynamic)}
\title{The Formation Operator Term}
\usage{
# binary: Form(formula)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}
}
\description{
The Formation Operator Term
}
\details{
This term accepts a model formula
and produces the corresponding model for the post-formation network:
effectively a network containing both previous time step's ties and ties just formed,
the union of the previous and current network. This is the equivalent of the
old-style \code{formation} model.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "Form", "subsection")}
}
\concept{durational}
\concept{operator}
