% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.tergm.R
\name{gof.tergm}
\alias{gof.tergm}
\alias{print.gof.tergm}
\alias{summary.gof.tergm}
\alias{plot.gof.tergm}
\title{Goodness-of-fit methods for TERGM CMLE and CMPLE fits}
\usage{
\method{gof}{tergm}(object, ...)

\method{print}{gof.tergm}(x, ...)

\method{summary}{gof.tergm}(object, ...)

\method{plot}{gof.tergm}(x, ..., main = "Goodness-of-fit diagnostics")
}
\arguments{
\item{object}{For \code{gof.tergm}, \code{\link{tergm}}
conditional MLE (CMLE) or conditional MPLE (CMPLE) fit. For the
others, a \code{gof.tergm} object returned by \code{gof.tergm}.}

\item{\dots}{Additional arguments passed through to the respective
functions in the \code{\link[=ergm-package]{ergm}} package.}

\item{x}{A \code{gof.tergm} object returned by
\code{gof.tergm}.}

\item{main}{Gives the title of the goodness-of-fit plot.}
}
\value{
For \code{gof.tergm}, an object of class
\code{gof.tergm}.

For the others, nothing.
}
\description{
For now, these are simple wrappers around \code{\link{gof.ergm}},
\code{\link{print.gof}}, \code{\link{summary.gof}}, and
\code{\link{plot.gof}}, respectively.
}
\examples{

data(samplk)

# Fit a transition from Time 1 to Time 2
samplk12 <- tergm(list(samplk1, samplk2) ~ Form(~edges+mutual+transitiveties+cyclicalties) +
                                           Diss(~edges+mutual+transitiveties+cyclicalties),
                                           estimate="CMLE")

samplk12.gof <- gof(samplk12)

samplk12.gof

summary(samplk12.gof)

plot(samplk12.gof)

plot(samplk12.gof, plotlogodds=TRUE)

}
\seealso{
\code{\link[=tergm]{tergm()}}, \code{\link[=ergm]{ergm()}}, \code{\link[=simulate.tergm]{simulate.tergm()}}, \code{\link[ergm:print.gof]{ergm::print.gof()}}, \code{\link[ergm:plot.gof]{ergm::plot.gof()}},
summary.gof, mcmc.diagnostics.ergm
}
\keyword{models}
