% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.netseries.R
\name{Lag-ergmTerm}
\alias{Lag-ergmTerm}
\alias{InitErgmTerm.Lag}
\title{An edge covariate on the previous network's statistics}
\usage{
# binary: Lag(
#           formula,
#           lag = 1,
#           transform = NULL,
#           lm = ~1,
#           subset = TRUE,
#           weights = 1,
#           contrasts = NULL,
#           offset = 0,
#           label = NULL
#         )
}
\arguments{
\item{formula}{a one-sided \code{\link[ergm:ergm]{ergm()}}-style formula with the terms to be evaluated}

\item{lag}{how many time steps to look back; at this time, only 1
is implemented.}

\item{transform}{a \verb{function(x, y)} given an array of covariates
and the sociomatrix of the \emph{previous} network to optionally
transform \code{x}, \code{NULL} to leave unchanged, or a character string
indicating a preset; current presets include \describe{

\item{\code{"signed"}}{multiplies each change statistic by \eqn{(2
y^{t-1}_{i,j} - 1)}}

\item{\code{"nonzero"}, "\code{positive}", "\code{negative}"}{indicator of whether
the change statistic has the respective value}

}}

\item{lm, subset, weights, contrasts, offset, label}{\strong{\code{\link[=NetSeries]{NetSeries()}} LHS only} arguments to specify time-varying parameters. See \code{\link[ergm.multi:N-ergmTerm-c9b59cc1]{N()}} term operator in the \pkg{ergm.multi} for details. \code{lm} formula may reference \code{.Time} for the network's time index, \code{.TimeID} for the its index in the network series (where the initial network is 1 and the first modelled network is 2), and \code{.TimeDelta} for the time elapsed between the network and the immediately previous network in the series.}
}
\description{
This term takes an ERGM formula that is evaluated on a
previous time step's network; the change statistics (optionally
transformed) are used as an edge covariate for the current
network. That is, a statistic of the form \eqn{\sum_{i,j}
  y^{t}_{i,j} \Delta_{i,j} g(y^{t-1})} or a transformed variant
\eqn{\sum_{i,j} y^{t}_{i,j} f(\Delta_{i,j} g(y^{t-1}),
  y^{t-1}_{i,j})}, where \eqn{Delta_{i,j} g(y) = g(y+(i,j)) -
  g(y-(i,j))}.
}
\references{
Almquist, Z. W. and Butts, C. T. (2014). Logistic Network
Regression for Scalable Analysis of Networks with Joint Edge/Vertex
Dynamics. \emph{Sociological Methodology}, 44(1),
273-321. \doi{10.1177/0081175013520159}
}
\seealso{
\code{\link[ergm:ergmTerm]{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "Lag", "subsection")}
}
\concept{dyad-independent}
\concept{operator}
\concept{temporal}
