% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.duration.R
\name{mean.age-ergmTerm}
\alias{mean.age-ergmTerm}
\alias{InitErgmTerm.mean.age}
\title{Average age of an extant tie}
\usage{
# binary: mean.age(emptyval=0, log=FALSE)
}
\arguments{
\item{emptyval}{can be used to specify the value returned if the network is empty. This is, technically, an arbitrary value, but it should
not have a substantial effect unless a non-negligible fraction of
networks at the parameter configuration of interest is empty.}

\item{log}{logical specifying if mean log age should be returned instead of mean age}
}
\description{
Average age of an extant tie
}
\details{
This term adds one statistic equaling the average, over all ties
present in the network, of the amount of time elapsed since
formation.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "mean.age", "subsection")}
}
\concept{durational}
