% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.EGMME.R
\name{Persist-ergmTerm}
\alias{Persist-ergmTerm}
\alias{InitErgmTerm.Persist (dynamic)}
\title{The Persistence Operator Term}
\usage{
# binary: Persist(formula)
}
\arguments{
\item{formula}{model formula}
}
\description{
The Persistence Operator Term
}
\details{
This term accepts a model formula
and produces the corresponding model for the
post-dissolution/persistence network: effectively the network containing
ties that persisted since the last time step.

This is the equivalent of the old-style \code{dissolution} model. So
a larger positive coefficient for \code{Persist()} operator means
less dissolution. It
produces the same results as the new \code{Diss()} operator, except the
signs of the coefficients are negated.
}
\concept{durational}
\concept{operator}
