% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.DynMLE.R
\name{staticDiscordTNT-ergmProposal}
\alias{staticDiscordTNT-ergmProposal}
\alias{InitErgmProposal.staticDiscordTNT}
\title{TNT proposal for CMLE fitting}
\description{
A version of \code{\link[ergm:TNT-ergmProposal]{TNT}} appropriate for CMLE fitting,
proposals stratified both by discordance status and edge status.  The argument
\code{ref} specifies the data relative to which discordance status is defined.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="staticDiscordTNT"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="staticDiscordTNT"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="staticDiscordTNT"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "staticDiscordTNT", "subsection")}
}
\keyword{internal}
