% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.EGMME.R
\name{.Statistics}
\alias{.Statistics}
\title{An Internal Operator Term for Subsetting Statistics}
\description{
The \code{.Statistics} term is used internally in \code{tergm}'s EGMME
to remove offset statistics from the targets model.  At this time it is not intended to
be used by the end user, but is documented here for completeness.  The behavior described
below may change without warning in the future, so do not rely on this term in your own code!

The \code{.Statistics} term takes two arguments: \code{formula} and \code{statistics}.  The
\code{formula} argument is a model formula and the \code{statistics} argument is a
\code{levels}-type argument for selecting statistics to retain in the model generated by
\code{formula} (and the input network).  If the original model is linear, so is the derived
model (meaning thetas are dropped and/or rearranged just as the etas are), and if the original
model is curved, so is the derived model, which retains the full set of original thetas, even
if some of them do not influence any retained etas.
}
