\name{stergm-internal}
\title{Internal stergm Objects}

\alias{InitConstraint.atleast}
\alias{InitConstraint.atmost}

\alias{InitErgmTerm.edge.ages}
\alias{InitErgmTerm.edgecov}
\alias{InitErgmTerm.edgecov.ages}
\alias{InitErgmTerm.edges.ageinterval}

\alias{InitMHP.dissolution}
\alias{InitMHP.dissolutionMLE}
\alias{InitMHP.dissolutionNonObservedMLE}
\alias{InitMHP.formation}
\alias{InitMHP.formationMLE}
\alias{InitMHP.formationMLETNT}
\alias{InitMHP.formationNonObservedMLE}
\alias{InitMHP.formationTNT}

\alias{control.godfather}
\alias{control.logLik.stergm}
\alias{ergm.godfather}
\alias{license.statnet}
\alias{statnetShowDoc}
\alias{statnetbrowseURL}
\alias{stergm.CMLE}
\alias{stergm.EGMME}
\alias{stergm.EGMME.GD}
\alias{stergm.EGMME.OS}
\alias{stergm.EGMME.SA}
\alias{stergm.EGMME.SA.Phase2.C}
\alias{stergm.EGMME.initialfit}
\alias{stergm.getMCMCsample}

\description{
Internal stergm functions.}

\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).}

\seealso{stergm, statnet-package}
\keyword{internal}
