% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stergm.utils.R
\name{.extract.fd.formulae}
\alias{.extract.fd.formulae}
\title{An Internal Function for Extracting (Some) Formation and Dissolution Formulas from a Combined Formula}
\usage{
.extract.fd.formulae(formula)
}
\arguments{
\item{formula}{a \code{formula}.}
}
\value{
A \code{list} containing \code{form}, \code{diss}, \code{nonsep}, and \code{all} formulas as described above.
}
\description{
This function is used in \code{tergm.EGMME.initialfit} and also when targets or monitoring
formulas are specified by characters.  It makes a basic attempt to identify the
formation and dissolution formulas within a larger combined formula (which may also
include nonseparable terms).  More specifically, terms in the combined formula beginning
with \code{Form} have their (first) argument extracted, and these arguments are accumulated into
an overall formation formula.  The environment of this overall formula is the environment
of the argument to \code{Form} when only one instance of \code{Form} occurs in the combined
formula; otherwise it is obtained by starting with the first \code{Form}'s argument's environment
and then updating the environment successively for each additional instance of \code{Form} using
\code{statnet.commmon}'s \code{nonsimp_update.formula} with \code{from.new = TRUE}.  The same procedure
is applied to instances of \code{Diss} in the combined formula to produce the dissolution formula.
Formulas containing non-separable terms in the overall formula and all terms in the formation, dissolution,
and non-separable formulas are also returned.  The non-separable formula's environment is
that of the combined formula, and the environment of the "all" formula is obtained by starting with the
first of formation, dissolution, and non-separable formulas that contains terms and proceding
successively through any others that do as well, using \code{statnet.commmon}'s \code{nonsimp_update.formula}
with \code{from.new = TRUE} as needed.

Signs outside of \code{Form} and \code{Diss} are propagated to the formation and dissolution formulas.
Instances of \code{Form} and \code{Diss} occurring inside other terms or inside of \code{offset} are
not recognized and will be interpreted as belonging to the nonseparable part of the combined
formula.  (Offsets can usually be propagated inside the \code{Form} or \code{Diss} in the original
formula specification to avoid this problem.)

If usage proves problematic, one may specify the monitoring and/or targets formulas explicitly
(rather than by characters), and one may pass initial coefficient values for the EGMME to avoid
running \code{tergm.EGMME.initialfit}.
}
