% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm-deprecated.R
\name{tergm-deprecated}
\alias{tergm-deprecated}
\title{Functions that will no longer be supported in future releases of the package}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons
}
\keyword{internal}
\keyword{misc}
