%  File man/tergm-internal.Rd in package tergm, part of the Statnet suite
%  of packages for network analysis, https://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  https://statnet.org/attribution
%
%  Copyright 2008-2020 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{stergm-internal}
\title{Internal tergm Objects}



%\alias{combine.networks}

%\alias{get.dev}

%\alias{network.extract.with.lasttoggle} 
%\alias{networkDynamic.apply.changes}
%\alias{all.same}
%\alias{to.networkDynamic.lasttoggle}


\alias{control.logLik.stergm}
%\alias{stergm.CMLE}
%\alias{stergm.EGMME}
%\alias{stergm.EGMME.GD}
%\alias{stergm.EGMME.SA}
%\alias{stergm.EGMME.SA.Phase2.C}
%\alias{stergm.EGMME.initialfit}
%\alias{stergm.getMCMCsample}

%\alias{is.lasttoggle}

\description{
Internal tergm functions.}

\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).}

\seealso{stergm, statnet-package}
\keyword{internal}
