% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm.godfather.R
\name{control.tergm.godfather}
\alias{control.tergm.godfather}
\title{Control parameters for \code{\link[=tergm.godfather]{tergm.godfather()}}.}
\usage{
control.tergm.godfather(term.options = NULL)
}
\arguments{
\item{term.options}{A list of additional arguments to be passed to term initializers. See \code{\link[=term.options]{? term.options}}.}
}
\description{
Returns a list of its arguments.
}
