% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm.godfather.R
\name{control.tergm.godfather}
\alias{control.tergm.godfather}
\title{Control parameters for \code{\link[=tergm.godfather]{tergm.godfather()}}.}
\usage{
control.tergm.godfather(term.options = NULL)
}
\arguments{
\item{term.options}{A list of optional settings such as calculation
tuning options to be passed to the \code{InitErgmTerm} functions.}
}
\description{
Returns a list of its arguments.
}
