% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.lasttoggle.R
\name{lasttoggle}
\alias{lasttoggle}
\title{Lasttoggle}
\description{
A data structure used by \code{tergm} for tracking of limited information
about dyad edge histories.
}
\details{
The \code{tergm} package handles durational information attached to
\code{\link[network:network]{network}} objects by way of the \code{time} and
\code{lasttoggle} network attributes.  The \code{lasttoggle} data
structure is a 3-column matrix; the first two columns are tails
and heads (respectively) of dyads, and the third column is the last
time at which the dyad was toggled.  The default last toggle time
is \code{-INT_MAX/2}.  Last toggle times for non-edges are
periodically cleared in the C code.  The \code{time} network
attribute is simply an integer, and together with the
\code{lasttoggle} data it determines the age of an extant
tie as \code{time + 1} minus the last toggle time for that dyad.
The default value for \code{time} is 0.
}
