% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.netseries.R
\name{NetSeries}
\alias{NetSeries}
\title{A network series representation.}
\usage{
NetSeries(...)
}
\arguments{
\item{...}{series specification, in one of three formats:
\enumerate{
\item A list of identically- dimensioned and directed networks.
\item Several networks as arguments.
\item A \code{\link{networkDynamic}} object and a numeric vector of time indices.
}}
}
\value{
A network object with temporal metadata.
}
\description{
A function for specifying the LHS of a temporal network series ERGM.
}
\examples{

data(samplk)

# Method 1: list of networks
monks <- NetSeries(list(samplk1,samplk2,samplk3))
ergm(monks ~ Form(~edges)+Diss(~edges))

# Method 2: networks as arguments
monks <- NetSeries(samplk1,samplk2,samplk3)
ergm(monks ~ Form(~edges)+Diss(~edges))

# Method 3: networkDynamic and time points:
## TODO

}
\seealso{
\link[=ergm-terms]{Help on model specification} for specific terms.
}
