% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm-deprecated.R, R/stergm.getMCMCsample.R
\name{tergm-deprecated}
\alias{tergm-deprecated}
\alias{stergm.getMCMCsample}
\alias{stergm.mcmcslave}
\title{Functions that will no longer be supported in future releases of the package}
\usage{
stergm.getMCMCsample(nw, model.form, model.diss, model.mon, proposal.form,
  proposal.diss, eta.form, eta.diss, control, verbose, ...)

stergm.mcmcslave(Clist.form, Clist.diss, Clist.mon, proposal.form,
  proposal.diss, eta.form, eta.diss, control, verbose)
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons
}
\section{Functions}{
\itemize{
\item \code{stergm.getMCMCsample}: Use \code{\link[=stergm_MCMC_sample]{stergm_MCMC_sample()}} instead.

\item \code{stergm.mcmcslave}: Use \code{\link[=stergm_MCMC_slave]{stergm_MCMC_slave()}} instead.
}}

\keyword{internal}
\keyword{misc}
