%  File man/tergm-internal.Rd in package tergm, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2014 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{stergm-internal}
\title{Internal tergm Objects}

\alias{InitMHP.dissolution}
\alias{InitMHP.dissolutionTNT}
\alias{InitMHP.dissolutionMLE}
\alias{InitMHP.dissolutionNonObservedMLE}
\alias{InitMHP.formation}
\alias{InitMHP.formationMLE}
\alias{InitMHP.formationMLETNT}
\alias{InitMHP.formationNonObservedMLE}
\alias{InitMHP.formationTNT}
\alias{InitMHP.dissolutionMLEblockdiag}
\alias{InitMHP.dissolutionNonObservedMLEblockdiag}
\alias{InitMHP.formationMLEblockdiag}
\alias{InitMHP.formationMLEblockdiagTNT}
\alias{InitMHP.formationNonObservedMLEblockdiag}
\alias{InitMHP.dissolutionMLETNT}
\alias{InitMHP.dissolutionMLEblockdiagTNT}

%\alias{combine.networks}

%\alias{get.dev}

%\alias{network.extract.with.lasttoggle} 
%\alias{networkDynamic.apply.changes}
%\alias{all.same}
%\alias{to.networkDynamic.lasttoggle}


\alias{control.logLik.stergm}
%\alias{stergm.CMLE}
%\alias{stergm.EGMME}
%\alias{stergm.EGMME.GD}
%\alias{stergm.EGMME.SA}
%\alias{stergm.EGMME.SA.Phase2.C}
%\alias{stergm.EGMME.initialfit}
%\alias{stergm.getMCMCsample}

%\alias{is.lasttoggle}

\description{
Internal tergm functions.}

\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).}

\seealso{stergm, statnet-package}
\keyword{internal}
