% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tergm.R, R/summary.tergm.R, R/tergm.R
\name{print.tergm}
\alias{print.tergm}
\alias{summary.tergm}
\alias{tergm}
\title{Temporal Exponential-Family Random Graph Models}
\usage{
\method{print}{tergm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{tergm}(object, ...)

tergm(
  formula,
  constraints = ~.,
  estimate,
  times = NULL,
  offset.coef = NULL,
  targets = NULL,
  target.stats = NULL,
  SAN.offsets = NULL,
  eval.loglik = NVL(getOption("tergm.eval.loglik"), getOption("ergm.eval.loglik")),
  control = control.tergm(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{\link{tergm}} object.}

\item{digits}{Significant digits for coefficients}

\item{\dots}{Additional arguments, to be passed to lower-level functions.}

\item{object}{A \code{tergm} object.}

\item{formula}{an ERGM formula.}

\item{constraints}{A one-sided formula specifying one or more constraints on
the support of the distribution of the networks being modeled, using syntax
similar to the \code{formula} argument. Multiple constraints may be given,
separated by \dQuote{+} operators.  Together with the model terms in the
formula and the reference measure, the constraints define the distribution
of networks being modeled.

The default is \code{~.}, for an unconstrained model.

See the \link[=ergm-constraints]{ERGM constraints} documentation for the
constraints implemented in the \strong{\link[=ergm-package]{ergm}} package.
Other packages may add their own constraints.

Note that not all possible combinations of constraints are supported.}

\item{estimate}{One of "EGMME" for Equilibrium Generalized Method of Moments
Estimation, based on a single network with some temporal information and
making an assumption that it is a product of a TERGM process running to its
stationary (equilibrium) distribution; "CMLE" for Conditional Maximum
Likelihood Estimation, modeling a transition between two networks, or
"CMPLE" for Conditional Maximum PseudoLikelihood Estimation, using MPLE
instead of MLE.  CMPLE is extremely inaccurate at this time.}

\item{times}{For CMLE and CMPLE estimation, times or indexes at
which the networks whose transition is to be modeled are
observed. Default to \code{c(0,1)} if \code{nw} is a
\code{\link[networkDynamic]{networkDynamic}} and to
\code{1:length(nw)} (all transitions) if \code{nw} is a
\code{\link{network.list}} or a \code{\link{list}}. Unused for
EGMME. Note that at this time, the selected time points will be
treated as temporally adjacent. Irregluarly spaced time series
are not supported at this time.}

\item{offset.coef}{Numeric vector to specify offset parameters.}

\item{targets}{One-sided \code{\link{ergm}}-style formula specifying
statistics whose moments are used for the EGMME. Unused for CMLE and CMPLE.
Targets is required for EGMME estimation. It may contain any valid ergm
terms. Any offset terms are used only during the preliminary SAN run;
they are removed automatically for the EGMME proper.}

\item{target.stats}{A vector specifying the values of the \code{targets}
statistics that EGMME will try to match.  Defaults to the statistics of
\code{nw}. Unused for CMLE and CMPLE.}

\item{SAN.offsets}{Offset coefficients (if any) to use during the SAN run.}

\item{eval.loglik}{Whether or not to calculate the log-likelihood
of a CMLE TERGM fit. See \code{\link{ergm}} for details. Can be
set globally via \code{option(tergm.eval.loglik=...)}, falling back to
\code{getOption("ergm.eval.loglik")} if not set.}

\item{control}{A list of control parameters for algorithm tuning.
Constructed using \code{\link{control.tergm}}.}

\item{verbose}{logical or integer; if TRUE or positive, the program will
print out progress information. Higher values result in more output.}
}
\value{
\code{\link{tergm}} returns an object of class \code{\link{tergm}}
that is a list consisting of the following elements:
\item{formula}{The model formula.}
\item{coef}{The fitted model coefficients.}
\item{targets}{For EGMME, the targets formula.}
\item{target.stats}{For EGMME, the target statistics.}
\item{estimate}{The type of estimate.}
\item{constraints}{The constraints used.}
\item{opt.history}{For EGMME, a matrix containing the full trace of
the optimization process: coefficients tried and target statistics simulated.}
\item{sample}{For EGMME, an \code{\link{mcmc}} object containing target
statistics sampled at the estimate.}
\item{covar}{For EGMME, the full estimated
variance-covariance matrix of the parameter estimates.}
\item{fit}{For CMLE and CMPLE, an \code{\link{ergm}} object from
fitting the model. For EGMME, stripped down \code{\link{ergm}}-like lists.}
\item{network}{For \code{estimate=="EGMME"}, the post-SAN network (or the original
network if SAN is not run); for \code{estimate=="CMLE"}
or \code{estimate=="CMPLE"}, a \code{\link{network.list}} (a discrete series
of networks) to which the model was fit.}
\item{control}{The control parameters used to fit the model.}

At this time, the \code{EGMME} \code{tergm} object also carries
some additional fit-related information (this may be changed in the future).

See the method \code{\link{print.tergm}} for details on how an
\code{\link{tergm}} object is printed.  Note that the method
\code{\link{summary.tergm}} returns a summary of the relevant parts of the
\code{\link{tergm}} object in concise summary format.
}
\description{
\code{\link{tergm}} is used for finding Temporal ERGMs' (TERGMs) and Separable Temporal ERGMs' (STERGMs)
Conditional MLE (CMLE) (Krivitsky and Handcock, 2010) and Equilibrium
Generalized Method of Moments Estimator (EGMME) (Krivitsky, 2009).
}
\details{
\strong{Model Terms} See \code{\link{ergm}} and \code{\link{ergm-terms}} for
details. At this time, only linear ERGM terms are allowed.  \itemize{
\item For a brief demonstration, please see the tergm package vignette:
\code{browseVignettes(package='tergm')} \item A more detailed tutorial is
available on the statnet wiki:
\url{https://statnet.org/Workshops/tergm_tutorial.html} }
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print the parameter estimates.

\item \code{summary}: Print the summary of the model fit.
}}

\examples{
\dontrun{
# EGMME Example
par(ask=FALSE)
n<-30
g0<-network.initialize(n,dir=FALSE)

#                     edges, degree(1), mean.age
target.stats<-c(      n*1/2,    n*0.6,        20)

dynfit<-tergm(g0 ~ Form(~edges + degree(1)) + Diss(~edges),
               targets = ~edges+degree(1)+mean.age,
               target.stats=target.stats, estimate="EGMME",
               control=control.tergm(SA.plot.progress=TRUE))

par(ask=TRUE)
mcmc.diagnostics(dynfit)
summary(dynfit)

# CMLE Example
data(samplk)

# Fit a transition from Time 1 to Time 2
samplk12 <- tergm(list(samplk1, samplk2)~
                  Form(~edges+mutual+transitiveties+cyclicalties)+
                  Diss(~edges+mutual+transitiveties+cyclicalties),
                  estimate="CMLE")

mcmc.diagnostics(samplk12)
summary(samplk12)

# Fit a transition from Time 1 to Time 2 and from Time 2 to Time 3 jointly
samplk123 <- tergm(list(samplk1, samplk2, samplk3)~
                   Form(~edges+mutual+transitiveties+cyclicalties)+
                   Diss(~edges+mutual+transitiveties+cyclicalties),
                   estimate="CMLE")

mcmc.diagnostics(samplk123)
summary(samplk123)
}
}
\references{
\itemize{

\item Krivitsky P.N. and Handcock M.S. (2014) A Separable Model for Dynamic Networks. \emph{Journal of the Royal Statistical Society, Series B}, 76(1): 29-46. \doi{10.1111/rssb.12014}

\item Krivitsky, P.N. (2012). Modeling of Dynamic Networks based on
Egocentric Data with Durational Information. \emph{Pennsylvania State
University Department of Statistics Technical Report}, 2012(2012-01).
\url{http://stat.psu.edu/research/technical-report-files/2012-technical-reports/modeling-of-dynamic-networks-based-on-egocentric-data-with-durational-information}

}
}
\seealso{
ergm, network, \\%v\\%, \\%n\\%, \code{\link{ergm-terms}}
}
