% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm.getMCMCsample.R
\name{tergm_MCMC_sample}
\alias{tergm_MCMC_sample}
\alias{tergm_MCMC_slave}
\title{Collects a sample of networks and returns the statistics of each sample}
\usage{
tergm_MCMC_sample(
  nw,
  model,
  model.mon = NULL,
  proposal,
  control,
  theta,
  verbose = FALSE,
  ...,
  eta = ergm.eta(theta, model$etamap)
)

tergm_MCMC_slave(Clist, proposal, eta, control, verbose)
}
\arguments{
\item{nw}{a \code{\link{network}} object}

\item{model}{the model, as returned by \code{\link{ergm_model}}}

\item{model.mon}{the optional monitoring model, as returned by \code{\link{ergm_model}}}

\item{proposal}{a list of parameters needed for
proposals}

\item{control}{list of control paramters, probably from
\code{\link{control.tergm}}}

\item{verbose}{logical; whether this and other functions should be verbose}

\item{eta}{vector of natural parameters.}

\item{Clist}{the "Clist" for the network and model under consideration,
as returned by \code{\link{ergm.Cprepare}}}
}
\value{
returns the MCMC sample as a list containing: \itemize{
\item statsmatrix: the matrix of sampled statistics for 'model'
RELATIVE TO INITIAL NETWORK \item newnetwork : the final network from the
sampling process \item changed : a toggle matrix, where the first column is
the timestamp of the toggle and the 2nd and 3rd columns are the head & tail
of the toggle; this is only returned if \code{control$changes} is not NULL
\item maxchanges : the "MCMC Dyn workspace"; see 'maxchanges' in the input
param list }
}
\description{
\code{tergm_MCMC_sample} is a low-level internal function not intended to
be called directly by end users. It collects a sample of networks and
returns the statistics of each sample, along with
a toggle matrix of the changes needed from the original network to each in
the sample.

\code{tergm_MCMC_slave} is an even
lower-level function that actually calls the C code.
}
\details{
This function is normally called inside \code{\link{simulate.tergm}} to
prepare inputs for the C sampling code and return its results
}
\seealso{
\code{\link{simulate.tergm}}
}
\keyword{internal}
